fixup! refactor: create adapter base-class and implementations for git and hg
