+++ lua/diffview/lib.lua
@@ -131,7 +131,8 @@ end
 
 ---@param range? { [1]: integer, [2]: integer }
 ---@param args string[]
-function M.file_history(range, args)
+function M.file_history(adapter, range, args)
+  -- TODO: Arg handling is git-specific too?
   local default_args = config.get_config().default_args.DiffviewFileHistory
   local argo = arg_parser.parse(vim.tbl_flatten({ default_args, args }))
   local paths = {}
@@ -142,134 +143,15 @@ function M.file_history(range, args)
     args,
   }), " "))
 
-  for _, path_arg in ipairs(argo.args) do
-    for _, path in ipairs(pl:vim_expand(path_arg, false, true)) do
-      local magic, pattern = vcs.pathspec_split(path)
-      pattern = pl:readlink(pattern) or pattern
-      table.insert(paths, magic .. pattern)
-    end
-  end
+  local log_options = adapter:file_history_options(range, args)
 
-  ---@type string
-  local cpath = argo:get_flag("C", { no_empty = true, expand = true })
-  local cfile = pl:vim_expand("%")
-  cfile = pl:readlink(cfile) or cfile
-
-  local top_indicators = {}
-  for _, path in ipairs(paths) do
-    if vcs.pathspec_split(path) == "" then
-      table.insert(top_indicators, pl:absolute(path, cpath))
-      break
-    end
-  end
-
-  table.insert(top_indicators, cpath and pl:realpath(cpath) or (
-      vim.bo.buftype == ""
-      and pl:absolute(cfile)
-      or nil
-    ))
-
-  if not cpath then
-    table.insert(top_indicators, pl:realpath("."))
-  end
-
-  local err, git_toplevel = M.find_git_toplevel(top_indicators)
-
-  if err then
-    utils.err(err)
-    return
-  end
-
-  ---@cast git_toplevel string
-  logger.lvl(1).s_debug(("Found git top-level: %s"):format(utils.str_quote(git_toplevel)))
-
-  rel_paths = vim.tbl_map(function(v)
-    return v == "." and "." or pl:relative(v, ".")
-  end, paths)
-
-  local cwd = cpath or vim.loop.cwd()
-  paths = vim.tbl_map(function(pathspec)
-    return vcs.pathspec_expand(git_toplevel, cwd, pathspec)
-  end, paths) --[[@as string[] ]]
-
-  ---@type string
-  local range_arg = argo:get_flag("range", { no_empty = true })
-  if range_arg then
-    local ok = vcs.verify_rev_arg(git_toplevel, range_arg)
-    if not ok then
-      utils.err(("Bad revision: %s"):format(utils.str_quote(range_arg)))
-      return
-    end
-
-    logger.lvl(1).s_debug(("Verified range rev: %s"):format(range_arg))
-  end
-
-  local log_flag_names = {
-    { "follow" },
-    { "first-parent" },
-    { "show-pulls" },
-    { "reflog" },
-    { "all" },
-    { "merges" },
-    { "no-merges" },
-    { "reverse" },
-    { "max-count", "n" },
-    { "L" },
-    { "diff-merges" },
-    { "author" },
-    { "grep" },
-    { "base" },
-  }
-
-  ---@type LogOptions
-  local log_options = { rev_range = range_arg }
-  for _, names in ipairs(log_flag_names) do
-    local key, _ = names[1]:gsub("%-", "_")
-    local v = argo:get_flag(names, {
-      expect_string = type(config.log_option_defaults[key]) ~= "boolean",
-      expect_list = names[1] == "L",
-    })
-    log_options[key] = v
-  end
-
-  if range then
-    paths, rel_paths = {}, {}
-    log_options.L = {
-      ("%d,%d:%s"):format(range[1], range[2], pl:relative(pl:absolute(cfile), git_toplevel))
-    }
-  end
-
-  log_options.path_args = paths
-
-  local ok, opt_description = vcs.file_history_dry_run(git_toplevel, log_options)
-
-  if not ok then
-    utils.info({
-      ("No git history for the target(s) given the current options! Targets: %s")
-        :format(#rel_paths == 0 and "':(top)'" or table.concat(vim.tbl_map(function(v)
-          return "'" .. v .. "'"
-        end, rel_paths) --[[@as vector ]], ", ")),
-      ("Current options: [ %s ]"):format(opt_description)
-    })
-    return
-  end
-
-  local git_ctx = {
-    toplevel = git_toplevel,
-    dir = vcs.git_dir(git_toplevel),
-  }
-
-  if not git_ctx.dir then
-    utils.err(
-      ("Failed to find the git dir for the repository: %s")
-      :format(utils.str_quote(git_ctx.toplevel))
-    )
-    return
+  if log_options == nil then
+    utils.err('Failed to create log options for file_history')
   end
 
   ---@type FileHistoryView
   local v = FileHistoryView({
-    git_ctx = git_ctx,
+    git_ctx = adapter,
     log_options = log_options,
   })
